Streamlined functions for loading MatrixMarket data.
Rather than having different functions for each compression type, we
just use a single function with the compression as an argument. This
reduces the number of separate functions (and documentation).

Also added an option for readers to guess the compression in the file
or buffer based on its leading magic numbers.