Require an explicit threads= argument in apply() and related functions.
This avoids relying on global variables (from OpenMP or for custom
schemes) and their associated problems, especially when dealing with
more complex applications where the nested parallelization could occur.

Refactored tests for the apply-based functions so that they explicitly
test for identical behavior in the multi-threaded case. Also extended
the custom parallelization test to check for identical results in all
apply-based functions, not just the sums.