Refactor for the v2 Matrix interface with more sophisticated workspaces. (#25)
This is a monster refactor that introduces block and index workspaces. The
block workspace allows extraction of a contiguous slice of each row or column,
while the indexed workspace allows extraction of a sorted and unique set of
indices from each row/column. This allows Matrix implementations to do more
appropriate precompute in the workspace, based on what the user actually wants
to extract, rather than doing this repeatedly in each row()/column() call. Our
assumption is that the block/index doesn't actually change across calls.

The Matrix interface now requires implementation of new_*_workspace() functions
that either accept nothing, a start/length pair, or a indexing vector. Each
row()/column() method now must accept a Workspace; these are overloaded to
determine whether to extract the full row/column or a subset thereof. As a
consequence, all uses of tatami must now instantiate a Workspace in order to
determine what to extract. This is probably for the best; high-perf
applications should be doing this anyway, and now we force them to do it.

The workspaces are also split into different classes by row and column. Each
row() or sparse_row()  method must accept a Row*Workspace, and vice versa for
the columns. This reduces the chance of errors involving construction and use
of a workspace of an inappropriate dimension in the row()/column() call. It
also provides more opportunities for optimization in Matrix implementations,
as different pre-compute can be done for row/column extraction.

Updated tests to reflect the use of these new workspaces. All tests now use
simulated data rather than the hard-coded matrices in data/. Also did some 
minor test fixes and improvements while I was here:

- Prevent divide-by-zero during division-on-left.
- Fleshed out the median tests to cover more scenarios.
- Added checks for sorting of indices in sparse_*() methods.
- Increase coverage of the dimension_preferences and other things.
- Fixed size calculations in HDF5 dense chunking.