Allow workspace caching for multiple passes over the same Matrix. (#33)
This involves adding a cache= option to all workspace methods, which allows
implementations to decide whether they want to allocate memory to cache some
intermediate values from the extraction of each row/column. Then, if the same
row/column is encountered during another pass through the matrix, the method
can just re-use the cached values rather than recomputing them. 

Currently, this is only useful for sliced or indexed queries on the primary
dimension of sparse matrices. With cache=true, we can cache the outcome of the
binary search to find the start of the slice/index of each row/column;
subsequent queries on the same row/column just re-use this cache. This should
improve performance of multi-pass algorithms include power iterations.