Further specialization of the workspaces according to dense/sparse extraction. (#34)
This introduces virtual workspace classes that are specific for the
dense/sparse extractor methods. Matrix implementations can now perform
dense/sparse-specific optimizations during workspace construction. We can also
eliminate the sparse_row() and sparse_column() methods in favor of just row()
and column() overloads that are distinguished according to the workspace class.

Another feature of sparse-specific workspaces is that the user can, during
workspace construction, specify whether to extract just the features, just the
values, both, or neither (i.e., just report the number of non-zero values).
This enables implementations to perform further optimizations. In such cases,
the invalid pointers in the SparseRange should be set to NULL.

We moved the default sparse overloads for row()/column() into its own
VirtualDenseMatrix class, which can be inherited from by various dense classes.
This avoids accidentally calling the default methods if we misspecify up an
overload for a sparse method, which would be a major performance regression.
Now, such errors can be caught at compile time, as a misspecified overload
would mean that the corresponding Matrix method is pure virtual.

Also, due to the increased complexity of the workspace class definitions, we
split the DelayedSubset file into one file per class for easier parsing.