Improved the transparency of apply'ing a function over a Matrix. (#43)
Previously, we hid all the dispatch details inside the apply() function, and
forced callers to create a horribly complicated factory class that could
accommodate the different dispatch calls. Now, we just ask users to handle
dispatch themselves, and provide some utilities and configuration classes to
make it easier to construct the appropriate Extractors. This should be easier
to use as the logic is not split between apply() and the user's factory.