Added LRU, oracle-based caching for the HDF5 dense matrix. (#46)
If an oracle is supplied, we cache the chunks corresponding to the next set of
future accesses. This is primarily helpful for reducing the number of entries
into the critical sections, as we can pre-fetch a number of chunks at once.
(Note that we still do separate calls to the HDF5 library in this section, to
avoid unnecessary reorganization of data in the hyperslab.) 

If an oracle is not supplied, we use an LRU cache to maximize the utility of 
previously loaded chunks. This is a qualified improvement over the previous
behavior of "just the next set of contiguous chunks", as now we avoid loading
irrelevant data for non-consecutive accesses. On the other hand, we're not
quite as efficient for consecutive accesses, especially as we need to re-enter
the critical section for each chunk - but if that's a concern, use an oracle.