Avoid doubling the memory usage for oracle-based caching in HDF5.
This adds a linear-time search at each iteration to find and recycle
unused cache_data elements for use in next_cache_data, rather than
resizing the elements of next_cache_data directly. This ensures that
cache memory usage stays below the user's specification. We already
have a linear-time step (w.r.t. number of chunks) when clearing the
contents of next_cache_exists, so it's not a real problem.