Support oracles for primary access of HDF5 sparse matrices. (#47)
This extracts into a contiguous buffer wherever possible, thus giving
near-optimal performance for consecutive accesses. If an oracle is not
provided, we use an LRU cache; this avoids loading more data than required.

This feature required the addition of a back() method in the OracleStream;
unlike the dense case, we can't predict the cache memory used by the next
element, as the number of non-zeros will vary. So we need to make the
prediction first, and if we're going to exceed the cache, we need to put the
prediction back into the stream for the next cycle.