Added a semi-compressed sparse matrix representation. (#48)
This provides a memory-efficient representation for sparse matrices of
non-negative integers for which the vast majority of entries are equal to 1.
One can imagine this to be an almost binary matrix, with some support for the
occasional larger counts. We implement this by only storing the indices,
duplicating them when the count is greater than 1.