Switch to perfect forwarding of variadic args for new_extractor.
This allows the user to pass the Options after the block/subset
arguments, which is more consistent with the Matrix method signatures.
It also preserves the const l-value reference to avoid a copy.

However, for internal use, we still keep the Options first, so that we
can operate on the Options object (e.g., to check the sparse extraction
parameters) while forwarding the remaining variadic arguments.