Consistent handling of const'ness in Delayed wrappers.
Now all make_* functions will only return non-const Matrix pointers,
even if they accept a const pointer. All make_* functions are templated
on Value_/Index_ types rather than an opaque Matrix_ type, which is more
self-documenting. (Note that this has some implications for the subset
block tests, which need to be updated to avoid type conflicts.)

A consequence of this is that we need to overload each make_* function
so that template argument deduction works correctly; otherwise, it fails
as the supplied pointer is not the same const'ness as the signature.

In addition, all delayed wrappers hold const Matrix pointers to indicate
that they are strictly read-only, to avoid any ambiguity. Similarly,
constructors should accept only const pointers, with the exception of
DelayedBind as it takes a vector of possibly non-const pointers.