Added the sparse_proportion and prefer_rows_proportion virtual methods.
The first method allows combined matrices (or wrappers around combined
matrices) to make more educated choices about whether they are sparse,
based on the sparsity status of the component submatrices.

prefer_rows_proportions() is a simplified dimension_preference() that is
now mandatory for all Matrices. This avoids subtle perf bugs when a
developer incorrectly defines the method in the subclass.