Consolidate secondary iteration code across sparse matrices. (#50)
This introduces a new set of utility classes to find the right indices during
iteration over the secondary dimension of compressed sparse matrices. The aim
is to avoid duplicating this logic over sparse matrix classes.

All classes now use the same implementation where it checks the immediate next
index (to optimize for consecutive iteration) before deciding to jump via a
binary search. We re-use the same approach when searching in reverse, though it
is a bit less clean as we do not cache the corresponding indices.

As it is a standalone class, we can also add dedicated tests for this
functionality, rather than reaching into the CSC's internals. On a similar
note, we tests some edge cases for semi-compressed secondary extraction.