Overhauled the secondary sparse extractor utility. (#51)
This adds an optimization to short-circuit the iteration across primary
dimensions during secondary extraction when there are no non-zero elements,
e.g., for a stretch of all-zero rows in a CSC matrix. It also improves the
caching of indices when decrementing for more efficient reverse iteration. 
This was achieved with some refactoring to move more logic out of the
various sparse Matrix classes and into the secondary workspace class.