Improved perf for reverse iteration in secondary sparse extractor. (#52)
This mostly involves using a dedicated "fail flag" of -1 for the index below
the current position when the position is already at the start of the primary
dimension. Doing so avoids a conditional to check for the validity of the index
before actually using the index itself. However, some work is required to deal
with unsigned integers when the -1 wraps around to the max value.

Optimization was based on timings from the test_sparse-priority-queue repo.