Overhauled isometric unary operations for performance. (#55)
One aspect of this refactor is to preserve the potential sparsity of the
operation, e.g., when adding/subtracting zeros. Conversely, when dividing by
zero, we indicate that sparsity is lost.

Another aspect of this refactor is to move the iteration over elements into the
operation class methods, rather than just asking for implementations of
element-wise operations. This should increase the opportunities for
vectorization, either automatically or with manual instructions.