Reorganized code for the sparse matrices. (#59)
This moves all the sparse-related code into its own top-level subdirectory, as
part of a general effort to preserve 'base' for the common interfaces. We also
move the sparse-related test code into its own subdirectory inside 'tests/src'.
Many other files need to be updated for these new header paths.

In preparation for the upcoming FragmentedSparseMatrix class, the primary
extraction code has been moved out of the CompressedSparseMatrix class. We
further use the PointerStorage_ template parameter to determine whether the
indices are concatenated (in which case we need explicit pointers a la CSC
matrices) or fragmented into one vector of indices per element of the primary
dimension (in which case we can just use the start/end as the limits).

Similarly, the secondary extractor class is refactored to account for the
distinction between concatenated and fragmented inputs. We also rename it (and
the associated file) to avoid confusion with the sparse matrix class.