Reorganize secondary sparse extraction code for minor perf, bugfixes.
The main change here is to move edge cases out of the loop, i.e., the handling
of the first or last element of the secondary dimension). This is now done once
outside the loop, which consolidates the if()'s into a single conditional.

We also remove the templating for the 'closest_current_index' short-circuits,
as this can be done in the loop body directly. This should reduce the number of
template instantiations and is a bit easier to read.

Finally, we fix a minor bug with the short circuits where 'last_request' was
not being updated when a short circuit was taken.