Improve handling of isometric expansion when collapsing sparsity. (#62)
If the underlying matrix is sparse, we extract sparse outputs first, apply the
operation, and then insert the results into the dense output buffer. This may
be preceded by a fill() for the constant result of the operation on a zero.
Doing so is faster than extracting dense output from a sparse matrix and then
applying the operation to each element of the dense vector (see the benchmark
repo at test-isometric_unary_sparse_expansion).