Turn Value_ into a class template parameter in all isometric helpers. (#63)
This allows each helper to use the Value_ type to perform checks during
construction. Currently, we use it to determine whether or not sparsity is
preserved, which is not trivial when Value_ supports non-finite values. 

The cost is that the DelayedIsometricOp's Value_ type now needs to be
coordinated with that of the helper. This is acceptable given that the
matrix's interface value type is constant in most applications.

The arithmetic operations benefit most from this change as we can now do more
stringent checks on whether sparsity is preserved, e.g., when multiplying by a
non-finite float value. We can also protect against division by zero when 
the environment does not support IEEE-754 floats.