Added a default <thread> implementation for tatami::parallelize. (#71)
This ensures that we get to parallelize even if OpenMP isn't available. Also
avoids the need for custom schemes in most (non-OpenMP enabled) applications.

Also added some error handling so that threads don't crash themselves but
propagate their errors to the main thread for re-throwing. Otherwise the
threads never get joined and that causes an application crash.

Overhauled the tests and actions to check OpenMP all the time (if available).
This cuts down on the amount of work in the CI.