Improve the user-friendliness of the no-copy overloads.
We export the tabulate_groups, total_groups helper functions to make it
easier to use the no-copy overload for grouped functions. We wrap them
in the stats:: namespace to avoid cluttering the main namespace, and we
update the docs on the overloads to point to these utilities.

The grouped medians now accept any vector-like object for the group
sizes. This makes it possible to, e.g., pass in ArrayView-wrapped
pointers rather than forcing people to pass in a std::vector.

We also simplified the no-copy overloads for the grouped means. These
don't actually need the actual number of observations per group, so they
now just take the total number of groups.