Complete overhaul of the extractor interfaces for more user-friendliness. (#84)
This changes almost everything, so much so that it would be easier to list
things that didn't change. The main change is that we provide an interface for
oracle-aware extractors, where the fetch() method doesn't require users to
supply an index that matches up with Oracle's prediction. This is more
ergonomic as it removes the possibility of any user errors.

It is also easier for Matrix developers, who can immediately specialize the
Extractor class for the oracle/no-oracle cases. This simplifies the
implementation of the Matrix extractors as the developer don't need to ensure
that the same extractor instance is valid before/after setting the oracle.

A consequence of this change is that now every Matrix has an oracle-aware
overload for each of the existing dense/sparse/row/column getters. To make life
easier for developers, we collapsed the row/column virtual methods into a
single method; now we don't have to write separate methods for *_row and
*_column, but instead can just pass a 'row=' argument to specify the dimension.

We also removed the DimensionSelectionType specialization on the extractors.
This simplifies the extractors to 4 variants - with or without an oracle (i.e.
oracular or myopic), and dense or sparse. Again, this makes life a bit easier
for implementers, who don't have to do so much crazy template magic to re-use
code across extractor variants. A side effect of this change is that each
extractor no longer stores the selection information, e.g., block start.  This
is deemed acceptable as the caller should have this info anyway.

The tests have been revamped to check every Matrix with oracle-aware access.
We now use parametrized tests to comprehensively check every access pattern.
We define static members in the fixture classes to reduce the overhead of each
test case, by just setting up the mock data once and re-using it. More tests
have also been added to account for correct handling of unsorted indices,
as well as to improve test coverage of various odds and ends.

Some individual Matrix classes also have some user-visible changes. The
semi-compressed sparse matrix has been removed, allowing us to greatly simplify
the secondary extraction code. Expectations for operation helpers in the
IsometricOp classes have been modified for greater generality. These two things
probably weren't used much, but it's a v2.0 bump anyway, so whatever.

Plus a bunch of code was moved around to different source files, and some
tidying up for the documentation.