Follow the rule of 5 for all interfaces to avoid confusion.
Currently it doesn't matter as (i) interfaces get defaulted copy constructors
that (ii) are used to substitute for move constructors and (iii) can be called
in the implicitly-defined move constructors of the derived class (see
https://stackoverflow.com/questions/29834274/abstract-classes-and-move-semantics).

Howwever, step (i) is deprecated as per C++11, and we might end up with compiler
warnings at some point.So we'll just default the rule of 5 just to be safe.

Also removed some unnecessary virtual destructors for the mock isometric helpers,
to avoid any confusion given that they aren't meant to be used polymorphically.