Shifted aspects of the isometric helper contracts back to runtime. (#99)
This aligns with the general policy of moving row-based parameters out of the
template - in this case, the margin_ for vector operations is now a regular
class member. This makes the isometric operation interfaces more flexible as
developers aren't forced into templating the helpers to meet the contract. 

Doing this means that we have to replace most of the static constexpr's with
runtime method calls. This shouldn't have any major perf impact as the
comparison of 'margin_' with 'row' is already done at runtime due to the latter
being a variable. Nonetheless, we allow the helper classes to skip this runtime
check if the relevant methods are not implemented, in which case they are
assumed to always return false (known at compile time via some SFINAE magic).

We do require a static constexpr to check whether the developer wants to use the
basic or advanced interface. This is made explicit for easier debugging, rather
than trying to guess the interface from the methods via SFINAE.