Enable value type transformations in the isometric unary operation. (#103)
This allows us to support operations that return a different type without
resorting to DelayedCast. For example, we can now directly return bool-ish
uint8's from a comparison between doubles, or we could also promote ints to
doubles during arithmetic operations, without an explicit cast at any point.

This change requires a modification to the unary isometric contract to accept
another pointer for storing the output in a potentially-different type. We
update all the existing helpers to accept this extra argument.

The only exception to this change is that of the substitution helpers, which
(by definition) must have the same output as the input matrix. After all, it is
performing a substitution into the underlying matrix, it's not doing a cast on
elements that don't pass the comparison and aren't being substituted.