Improved the modulo calculation so that it returns an integer type.
This makes better use of integer division if the inputs are integers,
instead of casting everything into a double for std::fmod.

A consequence of this change is that it creates a potential compile-time
divide-by-zero, so we now need to mark MODULO as potentially unsafe.

Also fixed the modulo for negative operands so that it is consistent
with the floored integer division. Added more tests accordingly.