Streamlined 2D strided access by computing the index in the loop body.
This replaces an offset that is updated across loop iterations. An index
computed once per loop is easier to reason about, and it also more
friendly to vectorization by removing dependencies between iterations.
Obviously, it also avoids pointer arithmetic and the associated UB
when updating the pointer location at the final loop iteration.

While this change involves an extra multiplication, the operation seems
to be eliminated by both Clang and GCC at a non-zero optimization level
so we don't have to take a performance hit - the only exception is when
the compiler decides to vectorize the inner loop, but the extra op would
be in service to greater overall speed, so that's acceptable.