Slap 'final' on all Extractor, Matrix and Oracle subclasses.
This provides some better opportunities for devirtualization. It also
prevents further inheritance from my subclasses - I don't know why a
developer would do so, and well, now they can't.

The exception is that of DenseMatrix and *SparseMatrix, which can't be
marked as final because we create our own subclasses of them for
back-compatibility. Maybe we can get rid of these in the next release.