Remove defaults for many template parameters.
This is most relevant for the Matrix interface types, as users should be
explicit about the type of Matrix that they're creating. For all we know,
double and int might not even be sufficient to represent the user's data.

We remove defaults for class templates that were introduced to avoid listing
all template parameters, given the lack of partial CTAD in C++17. These weren't
that useful unless you already knew what the default types were; otherwise
you'd end up with conflicts between the default and provided types. So now the
caller is expected to list out the types explicitly in the constructor.

We kept defaults for a few of the storage types in the conversion functions, as
these are transparent to the user and only provided for optimization purposes.
All back-compatibility functions also have defaults retained.

For now, we haven't touched the defaults for the helpers of the isometric
operations, as it is likely that these will need to be completely refactored.