Refactor isometric helpers to derive from a virtual interface. (#115)
This formalizes the interface, simplifies the implementation and avoids binary
bloat from instantiating multiple DelayedIsometricOperation classes for
different operations. The cost is that there is now some extra dispatch at each
step, but this should be minor - and can be avoided altogether by including the
specific helper class in the Operation's template parameters.

We also soft-deprecate the make_* functions to avoid confusing users with a
whole bunch of different options for construction. The cost is that people now
need to explicitly specify all the template parameters but that is probably
good practice anyway (and a good portion of the parameters for each Helper will
be validated during compilation by checking against the Operation's params).