Replace size_t's with autos or decltype where appropriate. (#117)
The aim is to use the most appropriate integer type, e.g., an arbitrary
container's size_type, which may not necessarily be a size_t. This is usually
safer as we don't have to worry about whether the limits on size_t are
applicable to the container. We also get rid of unnecessary casts to size_t. 

Any array accesses and associated pointer arithmetic are still performed with
size_t. We make an exception for std::vector::size_type, which we assume to be
size_t when used with the default allocator. All of these remaining uses of
size_t require std:: namespacing as well as inclusion of <cstddef>.