Protect against overflows when allocating containers via the sanisizer library. (#119)
We check the constructor and resize() calls to protect against implicit casts
to size_type that might cause overflow on some unusual systems. Most of the
time, Index_ is smaller than size_type, so these checks are optimized out;
moreover, we know that the dimensions must fit in a std::size_t (otherwise the
fetch() arrays wouldn't be allocatable) so the checks really only have an
effect if the container's size_type is smaller than a size_t.
 
More generally, while we're switching to sanisizer, we protect against overflow
in some cumulative sums, e.g., the +1 for the indptrs pointers in CSC/R
formats, the calculation of the total extent in DelayedBind and the non-unique
subset classes. We also use nd_offset() to compute the matrix offsets with
automatic casting. Some fiddling is required to avoid overflow in the oracles
when switching between Index_, the PredictionIndex, and size_t..

We don't bother to protect against overflow in reserve() statements, at least
for std::vector. The overflow itself is well-defined for unsigned integers, and
the consequences are not too important as reserve is only suggestive; the real
error will eventually come from a failure in the push_back().