Avoid sorting the primary indices in compress_sparse_triplets.
We don't need the primary indices to be sorted to construct a compressed sparse
matrix, so we skip it to avoid some unnecessary memory reads/writes.

We soft-deprecate the old compress_sparse_triplets() function and replace it
with an overload that only promises to sort the values + secondary indices.

We also add some more tests for the different reordering modes.