Slapped const on variables and function arguments for safety. (#120)
This reduces the risk of accidental modifications to read-only variables.

We wrap decltype() calls with the new I<> alias to allow creation of non-const
variables of the same type as a const variable, typically for loop counters.

We also replace raw pointers with references where possible, to eliminate the
possibility of NULL pointers where they shouldn't be present. This also avoids
the need to declare the pointer itself as const.

Also some minor tweaks to ensure that the correct headers are included, and to
perform some minor optimizations.